requirements.txt
typer[all]
requests
validators
cryptography
maxminddb-geolite2
python-whois